Create README.md
I'm an American living in Poland for 25 years.  I have a BS in CompSci/Math and an iMBA in International Business.  I have spent the last 25 years doing startups in Poland after roughly 5 years as a programmer in the U.S.  In 1997, I developed a Palm-top based solution for SFA doing data transfer over the Internet to a server farm developed by my company.  It was cloud adoption in the early days of cloud where getting companies trust in hosting their data was much more difficult :)  Back then I was on the cutting edge of tech.

I went into coffee in 2004, launching one of the first micro-roasters in Poland focused on roast-to-order for HoReCa.  After 12 years of that, sold my shares and returned to tech mostly in a PM and DevOps capacity.

In 2019, jumped into a startup to create smart home tech for the short-term rental sector.  We created a hardware controller to talk to smartlocks and proto-typed a smoking and noise sensor bundle.  Unfortunately, our seed funding got consumed during COVID with only minimal traction.  We were unable to quickly find next round financing.  So, I'm now taking some time off to jump back into tech to combine it with my business knowledge.  Soon, who knows that I'll get back into ? :) 

Hope to meet new people here and learn from the youngsters :)